refactor the data store
