+++ Federation/src/Domain/Models/IdentityClient.cs
@@ -2,7 +2,7 @@
 
 using Wangkanai.Domain;
 
-namespace Wangkanai.Identity;
+namespace Wangkanai.Federation.Models;
 
 public class IdentityClient : IdentityClient<string>
 {
@@ -29,7 +29,7 @@ public class IdentityClient<TKey> : IAuditable
 
 	public virtual TKey      Id           { get; set; } = default!;
 	public virtual string    ClientId     { get; set; }
-	public virtual string    ProtocolType { get; set; } = IdentityConstants.ProtocolTypes.OpenIdConnect;
+	public virtual string    ProtocolType { get; set; } = FederationConstants.ProtocolTypes.OpenIdConnect;
 	public virtual string    ClientName   { get; set; }
 	public virtual DateTime  Created      { get; set; } = DateTime.UtcNow;
 	public virtual DateTime? Updated      { get; set; }
