+++ Federation/src/Domain/Models/IdentityScope.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-namespace Wangkanai.Identity;
+namespace Wangkanai.Federation.Models;
 
 public class IdentityScope : IdentityScope<string> { }
