+++ Federation/src/Domain/Models/IdentityTenant.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-using Microsoft.AspNetCore.Identity;
-
-namespace Wangkanai.Identity;
+namespace Wangkanai.Federation.Models;
 
 public class IdentityTenant : IdentityTenant<string>
 {
