+++ Federation/src/Federation/DependencyInjection/Options/FederationOptions.cs
@@ -1,7 +1,5 @@
 ﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-using Microsoft.AspNetCore.Identity;
-
 namespace Wangkanai.Federation;
 
 /// <summary>
@@ -9,7 +7,7 @@ namespace Wangkanai.Federation;
 /// </summary>
 public sealed class FederationOptions
 {
-	public DiscoveryOptions      Discovery { get; set; } = new();
-	public EndpointsOptions      Endpoints { get; set; } = new();
-	public OperationStoreOptions Stores    { get; set; } = new();
+	public DiscoveryOptions       Discovery { get; set; } = new();
+	public EndpointsOptions       Endpoints { get; set; } = new();
+	public FederationStoreOptions Stores    { get; set; } = new();
 }
